+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/springdatajpa/QuestionSDJpaService.java
@@ -2,6 +2,7 @@ package ru.dpopkov.knowthenics.services.springdatajpa;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
+import ru.dpopkov.knowthenics.exceptions.data.NotFoundInRepositoryException;
 import ru.dpopkov.knowthenics.model.Question;
 import ru.dpopkov.knowthenics.repositories.QuestionRepository;
 import ru.dpopkov.knowthenics.services.QuestionService;
@@ -14,6 +15,15 @@ public class QuestionSDJpaService extends AbstractSDJpaService<Question> impleme
 
     public QuestionSDJpaService(QuestionRepository questionRepository) {
         super(questionRepository);
+    }
+
+    @Override
+    public Question findById(Long id) {
+        try {
+            return super.findById(id);
+        } catch (NotFoundInRepositoryException ex) {
+            throw new NotFoundInRepositoryException("Question not found for ID " + id, ex);
+        }
     }
 
     @Override