+++ knowthenics-web/src/test/java/ru/dpopkov/knowthenics/controllers/AnswerControllerTest.java
@@ -36,7 +36,9 @@ class AnswerControllerTest {
 
     @BeforeEach
     void setUp() {
-        mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
+        mockMvc = MockMvcBuilders.standaloneSetup(controller)
+                .setControllerAdvice(new ControllerExceptionHandler())
+                .build();
     }
 
     @Test